Generate New Data

Generates a new data frame (in the form of a tibble) with each variable held constant or varying as a unique ordered sequence. All possible unique combinations are included and the columns are in the same order as those in data with new columns added from ref.

Usage

new_data(
  data,
  seq = character(0),
  ref = list(),
  obs_only = list(character(0)),
  length_out = 30
)

Arguments

data

The data frame to generate the new data from.

seq

A character vector of the variables in data to generate sequences for or a named integer vector of the maximum length for each sequence where each name is the respective variable.

ref

A named list of vectors to add to the new data frame. The variables must not be in seq but they do not need to be in data.

obs_only

A character vector specifying the character, factor, and ordered (factor) variables in seq that should only be represented by their unique combinations or a list of character vectors specifying the groupings of character, factor and ordered (factor) variables in seq that should each only be represented by their unique combinations. If TRUE it is set to be all character, factor, and ordered (factor) variables specified in seq.

length_out

A count indicating the maximum length of sequences for all types of variables except character, factor and ordered factors. If the maximum length for each variable is specified in seq then length_out is ignored.

Value

A tibble of the new data.

Details

It is worth noting that ref can be used to specify reference values as well as sequences. The ref argument is useful for introducing new variables, overriding the default reference values, extrapolating outside the range of the data and/or changing the levels of a factor or time zone of a POSIXct vector. If an element of ref is a character vector and the corresponding column is a data frame, then the ref element is assigned the same factor levels as the column in the data. This is useful for choosing a factor level without having to set the correct levels.

See also

Examples

# an example data set
data <- tibble::tibble(
  vecint = c(1L, 3L),
  vecreal = c(1, 3),
  vecchar = c("b", "a"),
  vecdate = as.Date("2001-01-01", "2001-01-01")
)

# vary count while holding other values constant
new_data(data, "vecint")
#> # A tibble: 3 × 4
#>   vecint vecreal vecchar vecdate   
#>    <int>   <dbl> <chr>   <date>    
#> 1      1       2 a       2023-10-02
#> 2      2       2 a       2023-10-02
#> 3      3       2 a       2023-10-02
# vary continual
new_data(data, "vecreal")
#> # A tibble: 30 × 4
#>    vecint vecreal vecchar vecdate   
#>     <int>   <dbl> <chr>   <date>    
#>  1      2    1    a       2023-10-02
#>  2      2    1.07 a       2023-10-02
#>  3      2    1.14 a       2023-10-02
#>  4      2    1.21 a       2023-10-02
#>  5      2    1.28 a       2023-10-02
#>  6      2    1.34 a       2023-10-02
#>  7      2    1.41 a       2023-10-02
#>  8      2    1.48 a       2023-10-02
#>  9      2    1.55 a       2023-10-02
#> 10      2    1.62 a       2023-10-02
#> # ℹ 20 more rows
new_data(data, c("vecchar", "vecint"))
#> # A tibble: 6 × 4
#>   vecint vecreal vecchar vecdate   
#>    <int>   <dbl> <chr>   <date>    
#> 1      1       2 a       2023-10-02
#> 2      2       2 a       2023-10-02
#> 3      3       2 a       2023-10-02
#> 4      1       2 b       2023-10-02
#> 5      2       2 b       2023-10-02
#> 6      3       2 b       2023-10-02